develop some fix after review
